Fix integration test and docker build
