+++ test/testdata/check.golden
@@ -7,6 +7,7 @@ linkerd-api: can initialize the client.....................................[ok]
 linkerd-api: can query the control plane API...............................[ok]
 linkerd-api[kubernetes]: control plane can talk to Kubernetes..............[ok]
 linkerd-api[prometheus]: control plane can talk to Prometheus..............[ok]
+linkerd-api: no invalid service profiles...................................[ok]
 linkerd-version: can determine the latest version..........................[ok]
 linkerd-version: cli is up-to-date.........................................[ok]
 linkerd-version: control plane is up-to-date...............................[ok]
