+++ test/testdata/check.proxy.golden
@@ -7,6 +7,7 @@ linkerd-api: can initialize the client.....................................[ok]
 linkerd-api: can query the control plane API...............................[ok]
 linkerd-api[kubernetes]: control plane can talk to Kubernetes..............[ok]
 linkerd-api[prometheus]: control plane can talk to Prometheus..............[ok]
+linkerd-api: no invalid service profiles...................................[ok]
 linkerd-data-plane: data plane namespace exists............................[ok]
 linkerd-data-plane: data plane proxies are ready...........................[ok]
 linkerd-data-plane: data plane proxy metrics are present in Prometheus.....[ok]
